refactor: tenant tools in governance folder
